1-basic/4-micro-svc/requirements.txt
fastapi
fastapi[standard]
python-dotenv